commit github y eclipse por fin funcionando

+++ src/Backup.java
@@ -1,26 +1,15 @@
+package src;
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Calendar;
 
-public class Backup {
-
-	
+public class Backup { 
 
 	public String getFormatDay(int day) {
 		return (day > 10) ? "" + day : "0" + day;
 	}
-	private static Backup instance;
 
-	private Backup() {
-	}
-
-	public static Backup getInstance() {
-		if (instance == null) {
-			instance = new Backup();
-			
-		}
-		return instance;
-	}
 	public void doBackup() {
 		// Directorio Raiz de los Backups
 		File dirBackup = new File(Config.PATH_BACKUP);
@@ -38,13 +27,30 @@ public class Backup {
 		File origen = new File(Config.PATH_DB + Config.DBNAME);
 		format = "db-" + getFormatDay(cal.get(Calendar.DAY_OF_MONTH)) + ".db";
 		File destino = new File(dirBackup, format);
-		try{
-			Files.copy(origen.toPath(), destino.toPath());
-		}catch(Exception ex){
-			ex.printStackTrace();
+		
+		if (destino.exists()) {
+			return;
 		}
-		origen.renameTo(destino);
+		
+		try {
+			Files.copy(origen.toPath(), destino.toPath());
+		} catch (IOException e) { 
+			e.printStackTrace();
+			System.out.println(e);
+		}
+	}
+	
+	private static Backup instance;
+
+	private Backup() {
 	}
 
+	public static Backup getInstance() {
+		if (instance == null) {
+			instance = new Backup();
+		}
+		return instance;
+	}
+ 
 
 }
